pyproject.toml
[tool.poetry]
name = "1-channel-vae"
version = "0.1.0"
description = ""
authors = ["Saheed Faremi"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
shutils = "0.1.0"
toml = "^0.10.2"
matplotlib = "^3.8"
# torchvision = "^0.21.0"
wget = "^3.2"
scipy = "^1.15.1"
python = "^3.10"
torch = "^2.4.0"
gdown = "^5.2.0"
pydantic = "^2.8.2"
click = "^8.1.7"
pydantic-yaml = "^1.3.0"
scikit-learn = "1.6.1"
h5py = "3.12.1"
tensorboard = "^2.17.1"
torch-summary = "1.4.5"
PyYAML = "6.0.2"
mne = "1.9.0"
pandas = "2.2.3"
scikit-image = "0.25.2"
psutil = "7.0.0"
seaborn = "^0.13.2"
